Create translations module
