+++ packages/client/src/modules/charts/DetailsEnergyBars.tsx
@@ -9,6 +9,7 @@ import {
   Cell,
 } from "recharts";
 import { Persona } from "../persona/persona";
+import { translateConsumptionName } from "../translations";
 
 export { DetailsEnergyBars };
 
@@ -62,27 +63,4 @@ function DetailsEnergyBars({ persona }: { persona: Persona }) {
       </BarChart>
     </Card>
   );
-}
-
-function translateConsumptionName(value: string): string {
-  const translationByValue = {
-    airConditionning: "Air climatisé",
-    cleanCook: "Cuisine écologique",
-    brownGoods: "Produits bruns",
-    electricCar: "Voiture électrique",
-    light: "Lumière",
-    fossilCar: "Voiture thermique",
-    fossilHeating: "Chauffage thermique",
-    greyCar: "Voiture (gris)",
-    greyHouse: "Maison (gris)",
-    greyNumeric: "Numérique (gris)",
-    greyOther: "Autre (gris)",
-    greyTransport: "Transport (gris)",
-    food: "Nourriture",
-    noCarbonHeating: "Chauffage décarbonné",
-    plane: "Avion",
-    servicePublic: "Service public",
-    train: "Train",
-  };
-  return translationByValue[value as keyof typeof translationByValue] ?? value;
 }
